Cleaned up initialization code
